Made compatible with macOS mojave
After a 10hour-long fight I finally found how to make this work with macOS Mojave. Since apple deprecated stdlibc++, it is necessary to pass the extra argument "-stdlib=libc++" to both the compiler and the linker. Furthermore, as is the case in https://github.com/chokkan/crfsuite/issues/11, I modified the libraries given to the linker.